requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-ollama
langchain-mcp-adapters
langgraph-checkpoint-sqlite
streamlit
aiosqlite
fastmcp
python-dotenv
requests
faiss-cpu
duckduckgo-search